Answer the review: gate the cap refusal, split the answer suite, fix comment style
The eligibility gates now run before the quantity, so a corrupt cap only
refuses a cart the modifier would really apply to — a too-small cart never
reads it, and a new test pins that. The answer suites split in two under
the 400-line target, with the shared setup beside them. Two JSDoc comments
drop their semicolons and -ing forms for Simplified Technical English.
